[talos] Clean up scripts
  Move common code in common_hpp.py scripts in python module
  agimus_demos.talos.tools_hpp.
  Make scripts more robust to variations in robot model between
  talos full_v2 and Pyrene.
